backend/requirements.txt
fastapi
uvicorn
sqlalchemy
geoalchemy2
asyncpg
databases
psycopg2-binary
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
shapely
